вынес комментарии в отдельный файл
