+++ task6/index.html
@@ -6,7 +6,7 @@
     <title>Posts</title>
     <link rel="stylesheet" href="./css/normalize.css" />
     <link rel="stylesheet" href="./css/style.css" />
-    <script defer type="module" src="./js/index.js"></script>
+    <script defer type="module" src="./js/posts.js"></script>
   </head>
   <body>
     <div id="root"></div>
